Add files via upload
The goal of the project is to analyze the Customer Personality data to help businesses better understand their customers and allows them to develop targeted marketing campaigns according to the specific needs, behaviors, and concerns of different types of customers. This analysis can identify which customer segment is most likely to buy a product and then market the product only to that particular segment.